Pull all gcweb components into main layout file
