--- frontend/app/routes/_gcweb-app/language-switcher.tsx
@@ -1,30 +1,0 @@
-import { Link, type LinkProps, useLocation } from '@remix-run/react';
-
-import { useTranslation } from 'react-i18next';
-
-import { useClientEnv } from '~/components/client-env';
-
-/**
- * Props for the LanguageSwitcher component.
- */
-export type LanguageSwitcherProps = Omit<LinkProps, 'to' | 'reloadDocument'>;
-
-/**
- * A component that renders a link to switch between languages.
- */
-export function LanguageSwitcher({ ...props }: LanguageSwitcherProps) {
-  const langQueryParam = useClientEnv('LANG_QUERY_PARAM') ?? 'lang';
-  const { pathname } = useLocation();
-  const { i18n, t } = useTranslation(['gcweb']);
-
-  const to = `${pathname}?${langQueryParam}=${i18n.language === 'fr' ? 'en' : 'fr'}`;
-
-  return (
-    <Link {...props} to={to} data-testid="language-switcher" reloadDocument>
-      <span className="hidden-xs">{t('gcweb.language-switcher.alt-lang')}</span>
-      <abbr title="Français" className="visible-xs h3 mrgn-tp-sm mrgn-bttm-0 text-uppercase">
-        {t('gcweb.language-switcher.alt-lang-abbr')}
-      </abbr>
-    </Link>
-  );
-}
